static_cast: double will be better data type here.

+++ src/mongo/base/counter_test.cpp
@@ -53,7 +53,7 @@ TEST(CounterTest, Test1) {
     c.decrement(-1);
     ASSERT_EQUALS(c.get(), -1);
     c.increment();
-    ASSERT_EQUALS(static_cast<long long>(c), 0);
+    ASSERT_EQUALS(static_cast<double>(c), 0);
 }
 
 }  // namespace
